+++ packages/terra-status/src/terra-dev-site/doc/example/StatusDefault.jsx
@@ -6,7 +6,7 @@ const simpleText = <div style={{ padding: '5px' }}>Sample text</div>;
 
 const StatusDefault = () => (
   <div>
-    <Status color="#ff0000">{simpleText}</Status>
+    <Status color="#ff0000" visuallyHiddenText="Status High">{simpleText}</Status>
   </div>
 );