+++ packages/terra-status/CHANGELOG.md
@@ -4,6 +4,12 @@ Changelog
 Unreleased
 ----------
 
+2.29.0 - (November 9, 2018)
+------------------
+### Changed
+* Added VisuallyHiddenText prop to all examples in terra-status component.
+* Fixed lint errors
+
 2.28.0 - (October 24, 2018)
 ------------------
 ### Changed