+++ packages/terra-text/CHANGELOG.md
@@ -4,6 +4,15 @@ Changelog
 Unreleased
 ----------
 
+2.27.0 - (November 9, 2018)
+------------------
+### Added
+* Added additional font sizes
+* Added `300` option for weight prop
+
+### Changed
+* Fixed lint errors
+
 2.26.0 - (October 24, 2018)
 ------------------
 ### Changed